modules/panel-cbir/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "panel-cbir"
version = "0.1.0"
description = "Content-based image retrieval (CBIR) for scientific panel images."
readme = "README.md"
requires-python = ">=3.9"
license = {text = "MIT"}
authors = [{name = "Nivratti Boyane"}]

dependencies = [
  "numpy",
  "pillow",
  "loguru",
  "torch",
  "timm",
]

[project.optional-dependencies]
dev = ["pytest", "pyyaml", "tqdm"]

[project.scripts]
panel-cbir = "panel_cbir.cli:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
